etl/requirements_etl.txt
azure-core==1.32.0
azure-storage-blob==12.24.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
idna==3.10
isodate==0.7.2
psycopg2-binary==2.9.10
pycparser==2.22
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
python-dotenv==1.1.1
pendulum==3.1.0
